requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
qdrant-client
fastembed
flashrank
python-dotenv
pymupdf
pytesseract
pdf2image
Pillow
tqdm
streamlit
uvicorn